refactor(init): extract lazy dispatch tables to reduce __init__.py god modules (issue #1696)
The three largest __init__.py files (611, 350, 265 lines) accumulated bulk
because they inline large _LAZY_IMPORTS dispatch dicts and __all__ lists.
This refactor extracts those tables into dedicated _lazy_map.py modules,
bringing every __init__.py below 80 lines while preserving full backward
compatibility.

Changes:
- data_processor/core/__init__.py: 611 -> 61 lines; dispatch table -> _lazy_map.py
- model_generation/__init__.py: 350 -> 65 lines; dispatch table -> _lazy_map.py,
  convenience functions (quick_urdf, quick_build) -> _convenience.py
- signal_toolkit/__init__.py: 265 -> 77 lines; dispatch table -> _lazy_map.py

All public API imports remain backward compatible; no __all__ entries removed.
Architecture fitness test updated to accept dispatch table in _lazy_map.py.

Closes #1696
